+++ src/expand.cpp
@@ -38,6 +38,8 @@ static const char *toMessagePrefix(TildeExpandStatus status) {
     return "directory stack index out of range, up to directory stack limit ";
   case TildeExpandStatus::HAS_NULL:
     return "expanded path contains null characters: ";
+  case TildeExpandStatus::EMPTY_ASSIGN:
+    return "left-hand side of `=~' is empty";
   }
   assert(false);
   return "";
@@ -56,7 +58,9 @@ static void raiseTildeError(ARState &state, const DirStackProvider &provider,
     str += std::to_string(provider.size());
     str += "): ";
   }
-  str += ref;
+  if (status != TildeExpandStatus::EMPTY_ASSIGN) {
+    str += ref;
+  }
   raiseError(state, TYPE::TildeError, std::move(str));
 }
 
@@ -89,15 +93,37 @@ public:
   }
 };
 
-bool VM::applyTildeExpansion(ARState &state, StringRef path) {
-  std::string str = path.toString();
+static bool expandTildeWithLeftHandSide(ARState &state, const StringRef left, std::string &path) {
   DefaultDirStackProvider dirStackProvider(state);
-  if (const auto s = expandTilde(str, true, &dirStackProvider);
-      s != TildeExpandStatus::OK && state.has(RuntimeOption::FAIL_TILDE)) {
-    raiseTildeError(state, dirStackProvider, str, s);
+  TildeExpandStatus s;
+  if (left == "=") {
+    s = TildeExpandStatus::EMPTY_ASSIGN;
+  } else {
+    s = expandTilde(path, true, &dirStackProvider);
+  }
+  if (s != TildeExpandStatus::OK && state.has(RuntimeOption::FAIL_TILDE)) {
+    raiseTildeError(state, dirStackProvider, path, s);
     return false;
   }
-  state.stack.push(Value::createStr(std::move(str)));
+  return true;
+}
+
+bool VM::applyTildeExpansion(ARState &state, StringRef path, bool assign) {
+  std::string str = path.toString();
+  if (const StringRef left = assign ? state.stack.peek().asStrRef() : "";
+      !expandTildeWithLeftHandSide(state, left, str)) {
+    return false;
+  }
+  auto ret = Value::createStr(std::move(str));
+  if (assign) {
+    auto left = state.stack.pop();
+    const bool r = concatAsStr(state, left, ret, false);
+    if (r) {
+      state.stack.push(std::move(left));
+    }
+    return r;
+  }
+  state.stack.push(std::move(ret));
   return true;
 }
 
@@ -123,32 +149,75 @@ static void raiseGlobbingError(ARState &state, const GlobPatternWrapper &pattern
   raiseError(state, TYPE::GlobbingError, std::move(value));
 }
 
-static Value concatPath(ARState &state, const Value *const constPool, const Value *begin,
-                        const Value *end) {
-  auto ret = Value::createStr();
-  for (; begin != end; ++begin) {
-    if (begin->hasStrRef()) {
-      if (ret.appendAsStr(state, begin->asStrRef())) {
-        continue;
+struct TildeAssign {
+  unsigned int tildeCount{0};
+  unsigned int assignCount{0};
+};
+
+static bool isTildeExpansion(const Value *begin, const Value *cur, TildeAssign p) {
+  assert(cur->kind() == ValueKind::EXPAND_META);
+  if (cur == begin || ((cur - 1)->kind() == ValueKind::EXPAND_META &&
+                       (cur - 1)->asExpandMeta().first == ExpandMeta::ASSIGN)) {
+    const auto meta = cur->asExpandMeta().first;
+    return meta == ExpandMeta::TILDE && p.tildeCount == 0 && p.assignCount <= 1;
+  }
+  return false;
+}
+
+static Value concatPath(ARState &state, const Value *const constPool, const Value *const begin,
+                        const Value *const end) {
+  std::string out;
+  std::string tmp;
+  TildeAssign param;
+  bool tilde = false;
+  for (auto cur = begin; cur != end; ++cur) {
+    if (cur->hasStrRef()) {
+      if (!checkedAppend(cur->asStrRef(), StringObject::MAX_SIZE, tmp)) {
+        goto RAISE;
       }
-    } else if (begin->kind() == ValueKind::NUMBER) { // for escaped string
-      const StringRef ref = constPool[begin->asNum()].asStrRef();
-      std::string tmp;
-      const bool r = appendAsUnescaped(ref, StringObject::MAX_SIZE, tmp);
-      (void)r;
-      assert(r);
-      if (ret.appendAsStr(state, tmp)) {
-        continue;
+    } else if (cur->kind() == ValueKind::NUMBER) { // for escaped string
+      const StringRef ref = constPool[cur->asNum()].asStrRef();
+      if (!appendAsUnescaped(ref, StringObject::MAX_SIZE, tmp)) {
+        goto RAISE;
       }
     } else {
-      assert(begin->kind() == ValueKind::EXPAND_META);
-      if (ret.appendAsStr(state, begin->toString())) {
-        continue;
+      assert(cur->kind() == ValueKind::EXPAND_META);
+      if (isTildeExpansion(begin, cur, param)) {
+        tilde = true;
+        if (!checkedAppend(tmp, StringObject::MAX_SIZE, out)) {
+          goto RAISE;
+        }
+        tmp.clear();
+      }
+
+      if (const auto meta = cur->asExpandMeta().first; meta == ExpandMeta::TILDE) {
+        param.tildeCount++;
+      } else if (meta == ExpandMeta::ASSIGN) {
+        param.assignCount++;
+      }
+      if (!checkedAppend(cur->toString(), StringObject::MAX_SIZE, tmp)) {
+        goto RAISE;
       }
     }
-    return {};
   }
-  return ret;
+
+  if (tilde) {
+    assert(tmp[0] == '~');
+    if (!expandTildeWithLeftHandSide(state, out, tmp)) {
+      return {};
+    }
+  }
+  if (out.empty()) {
+    out = std::move(tmp);
+  } else if (!checkedAppend(tmp, StringObject::MAX_SIZE, out)) {
+    goto RAISE;
+  }
+  assert(out.size() <= StringObject::MAX_SIZE);
+  return Value::createStr(std::move(out));
+
+RAISE:
+  raiseStringLimit(state);
+  return {};
 }
 
 static Value joinAsPath(ARState &state, const GlobPatternWrapper &pattern) {
@@ -169,11 +238,15 @@ static Value joinAsPath(ARState &state, const GlobPatternWrapper &pattern) {
   return ret;
 }
 
-static bool concatAsGlobPattern(ARState &state, const Value *const constPool, const Value *begin,
-                                const Value *end, const bool tilde, GlobPatternWrapper &pattern) {
+static bool concatAsGlobPattern(ARState &state, const Value *const constPool,
+                                const Value *const begin, const Value *const end,
+                                GlobPatternWrapper &pattern) {
+  std::string prefix;
   std::string out;
-  for (; begin != end; ++begin) {
-    if (auto &v = *begin; v.hasStrRef()) {
+  bool tilde = false;
+  bool globMeta = false;
+  for (auto cur = begin; cur != end; ++cur) {
+    if (auto &v = *cur; v.hasStrRef()) {
       if (!appendAndEscapeGlobMeta(v.asStrRef(), StringObject::MAX_SIZE, out)) {
         goto NOMEM;
       }
@@ -184,25 +257,48 @@ static bool concatAsGlobPattern(ARState &state, const Value *const constPool, co
       }
     } else {
       assert(v.kind() == ValueKind::EXPAND_META);
+      const auto meta = v.asExpandMeta().first;
+      if (meta == ExpandMeta::TILDE && !tilde && (cur == begin || !prefix.empty())) {
+        tilde = true;
+      }
       if (!checkedAppend(v.toString(), StringObject::MAX_SIZE, out)) {
         goto NOMEM;
+      }
+      if (meta == ExpandMeta::ASSIGN && prefix.empty() && !globMeta && !tilde) {
+        prefix = out;
+        out.clear();
+      }
+      switch (meta) {
+      case ExpandMeta::ANY:
+      case ExpandMeta::ZERO_OR_MORE:
+      case ExpandMeta::BRACKET_OPEN:
+        globMeta = true;
+        break;
+      default:
+        break;
       }
     }
   }
   // check if glob path fragments have null character
-  if (const StringRef ref = out; unlikely(ref.hasNullChar())) {
+  if (unlikely(StringRef(out).hasNullChar() || StringRef(prefix).hasNullChar())) {
+    out.insert(0, prefix);
     raiseGlobbingErrorWithNull(state, out);
     return false;
   }
+
+  assert(!out.empty());
   pattern = GlobPatternWrapper::create(std::move(out));
   if (tilde && !pattern.getBaseDir().empty()) {
-    DefaultDirStackProvider provider(state);
-    if (const auto s = pattern.expandTilde(&provider);
-        s != TildeExpandStatus::OK && state.has(RuntimeOption::FAIL_TILDE)) {
-      raiseTildeError(state, provider, pattern.getBaseDir(), s);
-      return false;
+    assert(pattern.getBaseDir()[0] == '~');
+    if (!expandTildeWithLeftHandSide(state, prefix, pattern.refBaseDir())) {
+      return {};
     }
   }
+  if (const auto size = pattern.getBaseDir().size();
+      size > StringObject::MAX_SIZE || prefix.size() > StringObject::MAX_SIZE - size) {
+    goto NOMEM;
+  }
+  pattern.refBaseDir().insert(0, prefix);
   return true;
 
 NOMEM:
@@ -211,10 +307,10 @@ NOMEM:
 }
 
 bool VM::addGlobbingPath(ARState &state, ArrayObject &argv, const Value *const begin,
-                         const Value *const end, const bool tilde) {
+                         const Value *const end) {
   GlobPatternWrapper pattern;
   if (!concatAsGlobPattern(state, cast<CompiledCode>(state.stack.code())->getConstPool(), begin,
-                           end, tilde, pattern)) {
+                           end, pattern)) {
     return false;
   }
   if (pattern.getBaseDir().empty() && pattern.getPattern().size() == 1 &&
@@ -289,9 +385,15 @@ struct ExpandState {
 
 static bool needGlob(const Value *begin, const Value *end) {
   for (; begin != end; ++begin) {
-    if (const auto &v = *begin;
-        v.kind() == ValueKind::EXPAND_META && v.asExpandMeta().first != ExpandMeta::BRACKET_CLOSE) {
-      return true;
+    if (const auto &v = *begin; v.kind() == ValueKind::EXPAND_META) {
+      switch (v.asExpandMeta().first) {
+      case ExpandMeta::ANY:
+      case ExpandMeta::ZERO_OR_MORE:
+      case ExpandMeta::BRACKET_OPEN:
+        return true;
+      default:
+        break;
+      }
     }
   }
   return false;
@@ -317,7 +419,7 @@ static bool tryUpdate(int64_t &cur, const BaseObject &obj) {
 }
 
 bool VM::applyBraceExpansion(ARState &state, ArrayObject &argv, const Value *begin,
-                             const Value *end, ExpandOp expandOp) {
+                             const Value *end) {
   const unsigned int size = end - begin;
   assert(size <= UINT8_MAX);
   FlexBuffer<ExpandState> stack;
@@ -362,11 +464,6 @@ bool VM::applyBraceExpansion(ARState &state, ArrayObject &argv, const Value *beg
         i = iter->closeIndex;
         goto CONTINUE;
       }
-      case ExpandMeta::BRACE_TILDE:
-        if (usedSize) {
-          goto CONTINUE;
-        }
-        break;
       case ExpandMeta::BRACE_SEQ_OPEN: {
         i++;
         stack.push_back(ExpandState{
@@ -406,15 +503,8 @@ bool VM::applyBraceExpansion(ARState &state, ArrayObject &argv, const Value *beg
 
       const auto *vbegin = values.get();
       const auto *vend = vbegin + usedSize;
-      bool tilde = hasFlag(expandOp, ExpandOp::TILDE);
-      if (!tilde && usedSize > 0 && vbegin->kind() == ValueKind::EXPAND_META &&
-          vbegin->asExpandMeta().first == ExpandMeta::BRACE_TILDE) {
-        tilde = true;
-        ++vbegin; // skip meta
-      }
-
       if (needGlob(vbegin, vend)) {
-        if (!addGlobbingPath(state, argv, vbegin, vend, tilde)) {
+        if (!addGlobbingPath(state, argv, vbegin, vend)) {
           return false;
         }
       } else {
@@ -422,16 +512,6 @@ bool VM::applyBraceExpansion(ARState &state, ArrayObject &argv, const Value *beg
             concatPath(state, cast<CompiledCode>(state.stack.code())->getConstPool(), vbegin, vend);
         if (!path) {
           return false;
-        }
-        if (tilde) {
-          DefaultDirStackProvider dirStackProvider(state);
-          std::string str = path.asStrRef().toString();
-          if (const auto s = expandTilde(str, true, &dirStackProvider);
-              s != TildeExpandStatus::OK && state.has(RuntimeOption::FAIL_TILDE)) {
-            raiseTildeError(state, dirStackProvider, str, s);
-            return false;
-          }
-          path = Value::createStr(std::move(str));
         }
         if (!path.asStrRef().empty()) { // skip empty string
           if (!argv.append(state, std::move(path))) {
@@ -486,9 +566,9 @@ bool VM::addExpandingPath(ARState &state, const unsigned int size, ExpandOp expa
   const auto *end = &state.stack.peek();
   bool ret;
   if (hasFlag(expandOp, ExpandOp::BRACE)) {
-    ret = applyBraceExpansion(state, argv, begin, end, expandOp);
+    ret = applyBraceExpansion(state, argv, begin, end);
   } else {
-    ret = addGlobbingPath(state, argv, begin, end, hasFlag(expandOp, ExpandOp::TILDE));
+    ret = addGlobbingPath(state, argv, begin, end);
   }
 
   if (ret) { // pop operands