+++ src/node.cpp
@@ -431,15 +431,7 @@ void BinaryOpNode::dump(NodeDumper &dumper) const {
 
 void CmdArgNode::addSegmentNode(std::unique_ptr<Node> &&node) {
   this->updateToken(node->getToken());
-  if (isa<StringNode>(*node) && cast<StringNode>(*node).isTilde()) {
-    if (!this->segmentNodes.empty() && isa<WildCardNode>(*this->segmentNodes.back()) &&
-        cast<WildCardNode>(*this->segmentNodes.back()).isBraceMeta()) {
-      Token t = {node->getPos(), 0};
-      auto tilde = std::make_unique<WildCardNode>(t, ExpandMeta::BRACE_TILDE);
-      tilde->setExpand(false);
-      this->segmentNodes.push_back(std::move(tilde));
-    }
-  } else if (isa<WildCardNode>(*node)) {
+  if (isa<WildCardNode>(*node)) {
     auto &wildNode = cast<WildCardNode>(*node);
     if (wildNode.isExpand() && wildNode.isGlobMeta()) {
       setFlag(this->expansionAttr, GLOB);