+++ src/token_kind.h
@@ -159,6 +159,9 @@
   TOKEN(GLOB_ZERO_OR_MORE, "<glob *>")                                                             \
   TOKEN(GLOB_BRACKET_OPEN, "<glob [>")                                                             \
   TOKEN(GLOB_BRACKET_CLOSE, "<glob ]>")                                                            \
+  TOKEN(META_ASSIGN, "=")                                                                          \
+  TOKEN(META_COLON, ":")                                                                           \
+  TOKEN(TILDE, "~")                                                                                \
   TOKEN(BRACE_OPEN, "{")                                                                           \
   TOKEN(BRACE_CLOSE, "}")                                                                          \
   TOKEN(BRACE_SEP, ",")                                                                            \
@@ -348,6 +351,9 @@
 
 #define EACH_LA_cmdArg(OP)                                                                         \
   OP(CMD_ARG_PART)                                                                                 \
+  OP(META_ASSIGN)                                                                                  \
+  OP(META_COLON)                                                                                   \
+  OP(TILDE)                                                                                        \
   OP(BRACE_CHAR_SEQ)                                                                               \
   OP(BRACE_INT_SEQ)                                                                                \
   OP(GLOB_ANY)                                                                                     \